New Functionality
Created the new selector modal that lets u add open a modal with a picker of the option of the brands and the locations from the database, this in order to be able to make the update soon. Added a way to refresh the add screen for when u add a new location or brand to use be able to use it.